Deleted demo files/packages from springboot installation, renamed module and packages from demo to proper names
