+++ backend/src/main/java/TaskManager/TaskController.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package TaskManager;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
