+++ backend/src/main/java/TaskManager/TaskRepository.java
@@ -1,10 +1,6 @@
-package com.example.demo;
+package TaskManager;
 
-import jakarta.transaction.Transactional;
 import org.springframework.data.repository.CrudRepository;
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
 import java.util.List;
