+++ backend/src/test/java/com/Shafi/TaskManager/DemoApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.Shafi.TaskManager;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
